refactor: 패키지 경로 변경, 리턴 타입 변경 반영
